requirements.txt
Django==6.0.4
djangorestframework==3.17.1
coverage==7.13.5
